Map scrapping/requirements.txt
streamlit
easyocr
sentence-transformers
pandas
opencv-python
numpy
requests
Pillow
selenium
selenium-wire
webdriver-manager
blinker==1.7.0